+++ src/scripts/mod.rs
@@ -1,4 +1,5 @@
 mod detect;
+mod lang_mapping;
 mod script;
 
 pub use self::detect::detect_script;