+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add `Lang::values()` to iterate over all languages.
 * Add integration with `enum-map`
 * Implement `FromStr` for `Script` and `Lang`
+* Implement `Script::langs(&self) -> &[Lang]`
 
 ### v0.11.1 - 2020-11-28
 * Use Trigram tuple instead of heap allocated String. (~68% faster)